Reorganise the loadtests.
